// src/app/core/models/merchandise.model.ts
export class MerchandiseModel {
    _id?: string;
    name?: string;
    userName?: string;
    grabitValue?: number;
    userId?: string;
    desc?: string;
    email?: string;
    type?: string;
    city?: string;
    phone?: string;
    age?: string;
    attachments?: Array<any>;
    isVerified?: boolean;
    isLock?: boolean;
    isDeleted?: boolean;
    isAdmin?: boolean;
    profileUrl?: string;
    createdAt?: Date;
    updatedAt?: Date;
}